requirements.txt
anyio==4.12.0
APScheduler==3.11.1
cachetools==4.2.2
certifi==2025.11.12
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
setuptools==80.9.0
six==1.17.0
tornado==6.5.2
typing_extensions==4.15.0
tzlocal==5.3.1
